MOD : Separate feedbackSwitch
